use uint replace int8

+++ util/netutils/netutils.go
@@ -321,7 +321,7 @@ func (ar IPV4AddrRange) equals(ar2 IPV4AddrRange) bool {
 }
 
 func Masklen2Mask(maskLen int8) IPV4Addr {
-	return IPV4Addr(^(uint32(1<<(32-maskLen)) - 1))
+	return IPV4Addr(^(uint32(1<<(32-uint8(maskLen))) - 1))
 }
 
 type IPV4Prefix struct {
